Main Refactored
