+++ .idea/misc.xml
@@ -3,7 +3,8 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Android studio/Fitness_App/app/src/main/res/layout/activity_exercises.xml" value="0.12132725430597771" />
+        <entry key="..\:/Android studio/Fitness_App/app/src/main/res/layout/activity_all_exercises.xml" value="0.25" />
+        <entry key="..\:/Android studio/Fitness_App/app/src/main/res/layout/activity_exercises.xml" value="0.176" />
         <entry key="..\:/Android studio/Fitness_App/app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="..\:/Android studio/Fitness_App/app/src/main/res/layout/list_item.xml" value="0.12132725430597771" />
         <entry key="..\:/Users/Moham/Desktop/MyApplication/app/src/main/res/layout/activity_main.xml" value="0.17708333333333334" />
