+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication">
         <activity
+            android:name=".AllExercises"
+            android:exported="false" />
+        <activity
             android:name=".Exercises"
             android:exported="false" />
         <activity
