+++ app/src/main/res/layout/activity_all_exercises.xml
@@ -4,8 +4,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".AllExercises">
 
-
-
+    <ListView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:id="@+id/ListView"
+        tools:listitem="@layout/list_item"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
